Refactor main.py
Refactors main.py to separate state machine data from logic.

Splits all states into multiple sub-states, each either an entry point
asking a user to do some action or notifying them about some event, or
a handler, parsing and storing responses from the user. Because of the
way ConversationHandler is structured, only the latter are technically
states, and the former are implemented as simple functions being
called from the states.

Notion handling is moved into a separate function, which takes all the
data accumulated by the machine, renders it as a notion page and
submits that page to notion.